sdk: split query_verbs into module directory (9/9: extract summary)
Move summary, summary_report, and all summary_* helpers to
query_verbs/summary.rs.
Visibility: compute_summary, summary_subagent_session_filter,
collect_summary_agent_session_tree, collect_summary_connected_relationships,
aggregate_summary_relationship_stats, attribute_summary_cost_to_tools,
summary_turn_identity_key, summary_tool_attribution_method -> pub(crate);
SummaryRelationshipMatch, SummaryToolAgg (+ fields) -> pub(crate).
Add explicit imports for these pub(crate) items in tests.rs.
