+++ crates/relayburn-sdk/src/query_verbs/tests.rs
@@ -1,5 +1,11 @@
     use super::*;
     use super::flow::{bucket_subagents_per_turn, duration_to_since_iso};
+    use super::summary::{
+        aggregate_summary_relationship_stats, attribute_summary_cost_to_tools,
+        collect_summary_agent_session_tree, collect_summary_connected_relationships,
+        compute_summary, summary_subagent_session_filter, summary_tool_attribution_method,
+        summary_turn_identity_key, SummaryRelationshipMatch, SummaryToolAgg,
+    };
     use crate::reader::{
         RelationshipSourceKind, ToolCall, Usage, UserTurnBlock, UserTurnBlockKind,
     };
